Add: getLatLon function to hit TheZipCodes API
In order to allow the user to enter a zip code and have that zip codes
weather be displayed I used The ZipCodes API. I created a backend
function and route to obsfucate my API key and allow my front end to
call my baeckend server route `/lat-long-from-zip`. If successfull
an object will be returned with the correct lat lon pair.

*testing needs to be implemented for these functions/routes*