Refactor: combine zip/weather btns into 1 func
There was no longer a need to have a get weather and a get zip code
button in the user inteferface thus these were removed. The functions
that were called by these buttons are now called by the submit button
after a zip code is entered into the text input above it. This allows
users to input their zip code, click the submit and then the weather
data for that zipcode will be displayed to user.