Add: reduceWeatherTimeStamps function
The Open weather API currently returns 40 timstamps (8/day of week).
In order to make working with these days of the week easier I created
a reduced function that takes the API data and reduced it to 1x
timestamp per day of the week. The function also takes a days of the
week parameter to further refine the returned time stamps.