Add: functions and logic to display day of week weather details
In order to display each day of the weekes weather forcast in the DOM I
chose to dynamically create and fill DOM elements each time a weather
card is clicked on. I also added logic to continually update the
currently focused day of the week as well as which card is currently in
focus.