Add: logic for weather card expansion on focus / display of add details
In order to show details of the day of the week in focus I needed to
expand the width of the card as well as change the layout of the card
to prevent a large layout shift. I also added logic to show a spinner
on page load while waiting for the open weather API to return its data.